// convex/_generated/server.d.ts
type Env = {
  readonly CONVEX_CLOUD_URL: string;
  readonly CONVEX_SITE_URL: string;
  readonly CLERK_JWT_ISSUER_DOMAIN: string;
  readonly OPENAI_API_KEY: string;
};
// ...
export declare const action: ActionBuilder<DataModel, "public">;
// ...
export declare const env: Env;
// ...
export type ActionCtx = GenericActionCtx<DataModel>;